Move `GradlePropertiesController` and family to internal package
