+++ subprojects/core/src/main/java/org/gradle/initialization/ProjectPropertySettingBuildLoader.java
@@ -23,6 +23,7 @@ import org.gradle.api.internal.SettingsInternal;
 import org.gradle.api.internal.plugins.ExtraPropertiesExtensionInternal;
 import org.gradle.api.internal.project.ProjectInternal;
 import org.gradle.api.internal.properties.GradleProperties;
+import org.gradle.api.internal.properties.GradlePropertiesController;
 import org.gradle.initialization.properties.FilteringGradleProperties;
 
 import java.util.Set;
