+++ subprojects/core/src/main/java/org/gradle/internal/buildtree/BuildTreeScopeServices.java
@@ -48,9 +48,9 @@ import org.gradle.execution.TaskSelector;
 import org.gradle.execution.selection.BuildTaskSelector;
 import org.gradle.execution.selection.DefaultBuildTaskSelector;
 import org.gradle.initialization.BuildOptionBuildOperationProgressEventsEmitter;
-import org.gradle.initialization.DefaultGradlePropertiesController;
+import org.gradle.api.internal.properties.DefaultGradlePropertiesController;
 import org.gradle.initialization.Environment;
-import org.gradle.initialization.GradlePropertiesController;
+import org.gradle.api.internal.properties.GradlePropertiesController;
 import org.gradle.api.internal.properties.GradlePropertiesListener;
 import org.gradle.initialization.exception.DefaultExceptionAnalyser;
 import org.gradle.initialization.exception.ExceptionCollector;
